Rename wrappers to envs
