+++ malsim/envs/__init__.py
@@ -1,0 +1,10 @@
+from .malsim_vectorized_obs_env import MalSimVectorizedObsEnv
+from .gym_envs import AttackerEnv, DefenderEnv, register_envs
+
+# not needed, used to silence ruff F401
+__all__ = [
+    "MalSimVectorizedObsEnv",
+    "AttackerEnv",
+    "DefenderEnv",
+    "register_envs",
+]
