--- malsim/wrappers/__init__.py
@@ -1,2 +1,0 @@
-from .malsim_vectorized_obs_env import MalSimVectorizedObsEnv
-from .gym_wrapper import AttackerEnv, DefenderEnv, register_envs
