+++ tests/api_test.py
@@ -11,8 +11,8 @@ import gymnasium as gym
 from gymnasium.utils import env_checker
 from pettingzoo.test import parallel_api_test
 
-from malsim.wrappers import MalSimVectorizedObsEnv, AttackerEnv, DefenderEnv
-from malsim.wrappers.gym_wrapper import MaskingWrapper
+from malsim.envs import MalSimVectorizedObsEnv, AttackerEnv, DefenderEnv
+from malsim.envs.gym_envs import MaskingWrapper
 from malsim.agents.searchers import BreadthFirstAttacker, DepthFirstAttacker
 
 logger = logging.getLogger(__name__)
